face_attendance_project/requirements.txt
streamlit
opencv-python
face-recognition
face-recognition-models
numpy<2
pandas
pillow
qrcode
streamlit-js-eval
